[CI] Added Building Tests to EZGL CI
EZGL used to be tested as part of the VTR project. After being moved
into a sub-module, changes on the EZGL side need to be tested
independently. Added basic CI tests for building EZGL to ensure it at
least builds.

Removed a build folder that someone accidently pushed in 2 years ago.